fpat/paloalto_parameter_checker/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
paramiko==3.3.1
PyYAML==6.0.1
Jinja2==3.1.2
requests==2.31.0
python-dateutil==2.8.2
markupsafe==2.1.3